The PPD file support code from libucps was overtaken into libppd back
ing 2020, near 3 years ago:

https://openprinting.github.io/OpenPrinting-News-July-2020/#cups-filters

Even PPD files having been deprecated, there still happened a lot of
changes on the PPD support code in CUPS, some got merged into the copy
in libppd but not all.

So before the final release of libpps 2.0.0 I have looked through the
code and merged all fixes which did not make it into libppd yet.

Most are minor bug fixes, silencing warnings (especially of clang),
fixing typos in comments, coding style, ..., but there are also some
fixes for memory leaks, and even we add now some extra CUPS-specific
attributes to the PPD files we generate, to activate some of CUPS'
recently added new functionality, like providing tarnslations of
option and choice names without them being in the PPD files.

Here are some highlights:

- Add CUPS PPD attributes "*cupsLanguages: ...", "*cupsStringsURI ..."
  to generated PPDs so that CUPS loads printer-specific option
  names and translations from the printer and uses them without need
  of static translations in the PPD file.

- CUPS renames the PPD option choice name "Custom" to "_Custom" when a
  fixed choice is named as such, to distinguish from CUPS' facility
  for custom option values. Do the same when loading PPD files.

- Prevent duplicate PPD->IPP media-type name mappings

- When not specifying a media source and the page size is small
  (5x7" or smaller) do not request the photo tray but "auto" instead.

- Do not override color settings from print dialog (ColorModel) with
  print-color-mode setting.

- Make ppdFilterPSToPS() recognize "%%PageRequirements:" DSC comment.

- Correctly display "Xprinter" instead of Xerox for Xprinter devices

- Fix the job-pages-per-set value (used to apply finishings correctly)
  for duplex and N-up printing.

- Make the testppd test program also work if it is started from an
  environment with non-English locale.

+++ ppd/raster-interpret.c
@@ -55,7 +55,7 @@ typedef struct
     char	name[64];		// Name value
     double	number;			// Number value
     char	other[64];		// Other operator
-    char	string[64];		// Sring value
+    char	string[64];		// String value
   }			value;		// Value
 } _ppd_ps_obj_t;
 
@@ -131,7 +131,7 @@ ppdRasterInterpretPPD(
     cups_interpret_cb_t func)		// I - Optional page header callback
 					//     (@code NULL@ for none)
 {
-  int		status;			// Cummulative status
+  int		status;			// Cumulative status
   char		*code;			// Code to run
   const char	*val;			// Option value
   ppd_size_t	*size;			// Current size
@@ -355,10 +355,10 @@ ppdRasterInterpretPPD(
                                         h->cupsBorderlessScalingFactor);
   h->ImagingBoundingBox[3] = (unsigned)(top *
                                         h->cupsBorderlessScalingFactor);
-  h->cupsImagingBBox[0]    = (float)left;
-  h->cupsImagingBBox[1]    = (float)bottom;
-  h->cupsImagingBBox[2]    = (float)right;
-  h->cupsImagingBBox[3]    = (float)top;
+  h->cupsImagingBBox[0]    = left;
+  h->cupsImagingBBox[1]    = bottom;
+  h->cupsImagingBBox[2]    = right;
+  h->cupsImagingBBox[3]    = top;
 
   //
   // Use the callback to validate the page header...
